Add automation for tests, linter, and static analysis
Added support for automated tests, linter, and static analysis through GitHub Actions. This includes creating stubs for different action workflows and renaming them during the build process. The workflows are configured based on the user's selected testing framework (PHPUnit or Pest) and desired tools(Pint and PHPStan). A new kebabCase function was introduced to format the package name in README.md, which is also dynamically generated from a stub. These changes aim to facilitate and automate the package testing, linting, and static analysis processes, improving the workflow for users.