Add Spin Method for Installation Indication
A new `spin` methods was introduced to Prompts which I've added here to show some visual indication that the tool is running and installing everything it needs.

I also changed the ServiceProvider to not use Spatie's own package as I feel these are personal things one might wish to use and I don't want to force it.